Added documentation how to create a Custom-SAML-Assertion
To create a SAML-Assertion, only the filters we provide in the section Attributes can be used, but this is sometimes not flexible enough. 

For instance, when a User has been authenticated using an OAuth-Token, a new SAML-Assertion should be created on the fly. This is a quite complicated task for a very common scenario and we should help our customers by providing an example for it. 

BTW: My plan is to add more examples into the same GitHub repository and when appropriate link it into our documentation. I hope that is okay. I would like to share knowledge and avoid everybody is creating it's own implementation. Another example I'll tackle is this one: 
https://community.axway.com/s/question/0D52X00008NqmDPSAZ/signing-a-custom-jwt-with-rsa256